upgrade-project

+++ src/backend/Infrastructure/ImplementationContract/Services/ReviewService.cs
@@ -8,49 +8,93 @@ using Application.Extensions.ResultPattern;
 using Application.Filters;
 using Domain.Common;
 using Domain.Entities;
+using Domain.Enums;
 using Infrastructure.Extensions;
+using Microsoft.EntityFrameworkCore;
 
 namespace Infrastructure.ImplementationContract.Services;
 
-public class ReviewService(IReviewRepository repository) : IReviewService
+public class ReviewService(IReviewRepository repository,IBookingRepository bookingRepository) : IReviewService
 {
     public async Task<Result<PagedResponse<IEnumerable<ReviewReadInfo>>>> GetAllAsync(ReviewFilter filter)
     {
-        return await Task.Run(() =>
-        {
             Expression<Func<Review, bool>> filterExpression = spec =>
                 (filter.UserId == null || spec.UserId == filter.UserId) &&
                 (filter.CarId == null || spec.CarId == filter.CarId) &&
-                (filter.Rating == null || spec.Rating == filter.Rating);
-
+                (filter.Rating == null || spec.Rating == filter.Rating) &&
+                (filter.FromDate == null || spec.CreatedAt >= filter.FromDate) &&
+                (filter.ToDate == null || spec.CreatedAt <= filter.ToDate);
+            
+            
             Result<IQueryable<Review>> request = repository.Find(filterExpression);
 
             if (!request.IsSuccess)
                 return Result<PagedResponse<IEnumerable<ReviewReadInfo>>>.Failure(request.Error);
 
-            List<ReviewReadInfo> query = request.Value!.Select(x => x.ToRead()).ToList();
+            var query = request.Value!.AsNoTracking();
+            
+            int count = await query.CountAsync();
 
-            int count = query.Count;
-
-            IEnumerable<ReviewReadInfo> spec = query.Page(filter.PageNumber, filter.PageSize);
+            IEnumerable<ReviewReadInfo> data = await query
+                .OrderByDescending(x=>x.CreatedAt)
+                .Skip((filter.PageNumber-1)*filter.PageSize)
+                .Take(filter.PageSize)
+                .Select(x=>x.ToRead())
+                .ToListAsync();
 
             PagedResponse<IEnumerable<ReviewReadInfo>> res =
-                PagedResponse<IEnumerable<ReviewReadInfo>>.Create(filter.PageNumber, filter.PageSize, count, spec);
+                PagedResponse<IEnumerable<ReviewReadInfo>>.Create(filter.PageNumber, filter.PageSize, count, data);
 
             return Result<PagedResponse<IEnumerable<ReviewReadInfo>>>.Success(res);
-        });
+    }
+
+    public async Task<Result<IEnumerable<ReviewReadInfo>>> GetByCarIdAsync(int carId)
+    {
+        var res = repository.Find(x => x.CarId == carId);
+
+        if (!res.IsSuccess)
+            return Result<IEnumerable<ReviewReadInfo>>.Failure(res.Error);
+
+        var data = await res.Value!
+            .AsNoTracking()
+            .OrderByDescending(x=>x.CreatedAt)
+            .Select(x=>x.ToRead())
+            .ToListAsync();
+
+        return Result<IEnumerable<ReviewReadInfo>>.Success(data);
     }
 
     public async Task<Result<ReviewReadInfo>> GetByIdAsync(int id)
     {
         Result<Review?> res = await repository.GetByIdAsync(id);
-        if (!res.IsSuccess) return Result<ReviewReadInfo>.Failure(res.Error);
+        
+        if (!res.IsSuccess || res.Value is null) return Result<ReviewReadInfo>.Failure(Error.NotFound("Review not found"));
 
-        return Result<ReviewReadInfo>.Success(res.Value!.ToRead());
+        return Result<ReviewReadInfo>.Success(res.Value.ToRead());
     }
 
     public async Task<BaseResult> CreateAsync(int userId,ReviewCreateInfo createInfo)
     {
+        var hasCompletedBooking = bookingRepository
+            .Find(b => b.UserId == userId &&
+                       b.CarId == createInfo.CarId &&
+                       b.BookingStatus == BookingStatus.Completed);
+
+        if (!hasCompletedBooking.IsSuccess ||
+            !await hasCompletedBooking.Value!.AnyAsync())
+        {
+            return BaseResult.Failure(Error.BadRequest("You can leave review only after completed booking"));
+        }
+        
+        var exists = repository.Find(x =>
+            x.UserId == userId &&
+            x.CarId == createInfo.CarId);
+      
+
+        if (exists.IsSuccess && await exists.Value!.AnyAsync())
+            return BaseResult.Failure(
+                Error.Conflict("You already reviewed this car."));
+        
         Result<int> res = await repository.AddAsync(createInfo.ToEntity(userId));
 
         return res.IsSuccess
@@ -62,7 +106,7 @@ public class ReviewService(IReviewRepository repository) : IReviewService
     {
         Result<Review?> res = await repository.GetByIdAsync(id);
 
-        if (!res.IsSuccess) return BaseResult.Failure(Error.NotFound());
+        if (!res.IsSuccess || res.Value is null) return BaseResult.Failure(Error.NotFound("Review not found"));
 
         Result<int> result = await repository.UpdateAsync(res.Value!.ToEntity(updateInfo));
 
@@ -74,7 +118,7 @@ public class ReviewService(IReviewRepository repository) : IReviewService
     public async Task<BaseResult> DeleteAsync(int id)
     {
         Result<Review?> res = await repository.GetByIdAsync(id);
-        if (!res.IsSuccess) return BaseResult.Failure(Error.NotFound());
+        if (!res.IsSuccess || res.Value is null) return BaseResult.Failure(Error.NotFound("Review not found"));
 
         Result<int> result = await repository.DeleteAsync(id);
